requirements.txt
django
git+https://github.com/lememilitar/lemeauthpy.git
pytest-django
